+++ arm9/arm9.lsf
@@ -317,7 +317,6 @@ Static arm9
 	Object unk_0208A258.o
 	Object unk_0208A300.o
 	Object unk_0208A338.o
-	Object unk_0208A998_c.o
 	Object unk_0208A998.o
 	### Third-party libraries ###
 	Object unk_0208AC14.o
