--- arm9/asm/unk_0208A998.s
@@ -1,13 +1,0 @@
-    .include "asm/macros.inc"
-    .include "global.inc"
-
-	.extern gMain
-
-	.section .rodata
-
-
-	.section .bss
-	.text
-
-
-
